Initial check-in.
This file is for developers and is not distributed in the release files.

If you are a developer and you want to work directly from CVS, copy this file to
defines.php and edit the database DSN handle so it will work with your database setup.

This file should be updated when the installer is updated in case any new constants
are created.