The build.xml file was reworked and should now be building a correct build.
The INSTALL was updated with more information that outlines the install process.

install.php was modified to adhere to what the build.xml and the INSTALL documentation
is.

KNOWN BUG - the build process with ant is changing the images somehow so when you
open the web site, all of the images are broken.